requirements.txt
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
dnspython==2.7.0
eventlet==0.39.1
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
lxml==5.3.2
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
opencv-python==4.11.0.86
oscrypto==1.3.0
pillow==11.1.0
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
python-bidi==0.6.6
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
qrcode==8.1
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
setuptools==78.1.0
simple-websocket==1.1.0
six==1.17.0
svglib==1.5.1
sympy==1.13.1
tiktoken==0.9.0
tinycss2==1.4.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.2.0
xhtml2pdf==0.2.17
